// frontend/src/apiService/models/TopicExtended.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AnswerPublicExtended } from './AnswerPublicExtended';
import {
    AnswerPublicExtendedFromJSON,
    AnswerPublicExtendedFromJSONTyped,
    AnswerPublicExtendedToJSON,
    AnswerPublicExtendedToJSONTyped,
} from './AnswerPublicExtended';
import type { Summary } from './Summary';
import {
    SummaryFromJSON,
    SummaryFromJSONTyped,
    SummaryToJSON,
    SummaryToJSONTyped,
} from './Summary';

/**
 * 
 * @export
 * @interface TopicExtended
 */
export interface TopicExtended {
    /**
     * 
     * @type {string}
     * @memberof TopicExtended
     */
    algorithm: string;
    /**
     * 
     * @type {string}
     * @memberof TopicExtended
     */
    questionId: string;
    /**
     * 
     * @type {string}
     * @memberof TopicExtended
     */
    label: string;
    /**
     * 
     * @type {string}
     * @memberof TopicExtended
     */
    topic: string;
    /**
     * 
     * @type {number}
     * @memberof TopicExtended
     */
    score?: number;
    /**
     * 
     * @type {string}
     * @memberof TopicExtended
     */
    id: string;
    /**
     * 
     * @type {Summary}
     * @memberof TopicExtended
     */
    summary: Summary | null;
    /**
     * 
     * @type {Array<AnswerPublicExtended>}
     * @memberof TopicExtended
     */
    answers: Array<AnswerPublicExtended>;
}

/**
 * Check if a given object implements the TopicExtended interface.
 */
export function instanceOfTopicExtended(value: object): value is TopicExtended {
    if (!('algorithm' in value) || value['algorithm'] === undefined) return false;
    if (!('questionId' in value) || value['questionId'] === undefined) return false;
    if (!('label' in value) || value['label'] === undefined) return false;
    if (!('topic' in value) || value['topic'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('summary' in value) || value['summary'] === undefined) return false;
    if (!('answers' in value) || value['answers'] === undefined) return false;
    return true;
}

export function TopicExtendedFromJSON(json: any): TopicExtended {
    return TopicExtendedFromJSONTyped(json, false);
}

export function TopicExtendedFromJSONTyped(json: any, ignoreDiscriminator: boolean): TopicExtended {
    if (json == null) {
        return json;
    }
    return {
        
        'algorithm': json['algorithm'],
        'questionId': json['question_id'],
        'label': json['label'],
        'topic': json['topic'],
        'score': json['score'] == null ? undefined : json['score'],
        'id': json['id'],
        'summary': SummaryFromJSON(json['summary']),
        'answers': ((json['answers'] as Array<any>).map(AnswerPublicExtendedFromJSON)),
    };
}

export function TopicExtendedToJSON(json: any): TopicExtended {
    return TopicExtendedToJSONTyped(json, false);
}

export function TopicExtendedToJSONTyped(value?: TopicExtended | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'algorithm': value['algorithm'],
        'question_id': value['questionId'],
        'label': value['label'],
        'topic': value['topic'],
        'score': value['score'],
        'id': value['id'],
        'summary': SummaryToJSON(value['summary']),
        'answers': ((value['answers'] as Array<any>).map(AnswerPublicExtendedToJSON)),
    };
}